Points now integrated, Asteroids now move
There were occurring issues with the points where the points were not integrated into the game and was left out but now issue is resolved and points are now integrated and now display, even on the end screen game. Asteroids now spawn in and when shot at give you points depending on the asteroid and will spawn in new asteroids.

Project: Amazonas